requirements.txt
# Flask and Extensions
Flask==2.0.1
Werkzeug==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10

# MongoDB Support
Flask-MongoEngine==1.0.0
mongoengine==0.24.2
pymongo==4.3.3

# Database
SQLAlchemy==1.4.41
alembic==1.8.1
psycopg2-binary==2.9.9

# Security
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
cryptography==41.0.5

# Forms and Validation
WTForms==3.0.1
email-validator==2.0.0.post2

# Date/Time
pytz==2023.3
python-dateutil==2.8.2

# CLI
click==8.1.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0

# Development Tools
black==23.7.0
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0

# Monitoring and Logging
sentry-sdk[flask]==1.29.2

# Task Queue (for background jobs)
celery==5.3.1
redis==4.6.0

# API Documentation
flasgger==0.9.7.1
apispec==6.3.0

# Utilities
Pillow==10.0.0
requests==2.31.0
python-slugify==8.0.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Natural Language Processing
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.25.2
pandas==2.0.3
